+++ CHANGELOG.md
@@ -11,7 +11,10 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
-- None
+- [#1416](https://github.com/paper-trail-gem/paper_trail/pull/1416) - Adds a
+  model-configurable option `synchronize_version_creation_timestamp` which, if
+  set to false, opts out of synchronizing timestamps between `Version.created_at`
+  and the record's `updated_at`.
 
 ### Fixed